simplify boostrap script by doing in-place edits
This whole bootstrap script is kind of a hack. In the long term I think
a proper makefile is in order along with some more conventional
packaging routines, but for now at least it helps in some cases.
Reading it on the other hand was kind of painful. As long as you are
clobbering the original script with an edited version anyway there is no
advantage of hacking it in rather than just doing an in-place edit.