expand section on testing with example output
Knowing what the console output should look like it everything worked
properly can be quite useful to the first time user.

Note this section will need fixing when proper distribution packages
become available with the location on a file to test with. Currently the
assumption is that the user is sitting in the repository directory
already which might be fine for people who built form source themselves
but that isn't a safe assumption for a packaged release.