check build conditions before blindly using `git submodule`
This should allow the bootstrap script to be run on both packaged source
distributions as well as from in a git repository. The autoreconf
routine is not strictly necessary in the former situation but it is
recommended procedure for building on most distros. Checking the
conditions in the bootstrap script and handling the case of being in a
git repository but not having the submodule initialized should allow the
same build instructions to apply to all scenarios, simplifying the
documentation and reducing confusion.