reduce amount of fill content to test one thing at a time
The original Minion Pro font for the chapter break might have not shown
this problem, but on systems without that font the fallback might have a
different X height. On my system this test would wrap pages, which
introduced a new problem...that may need a test of it's own, but in the
spirit of testing one thing at a time this reduces the total content to
be on one page even if the selected font isn't around.