make dependency on luafilesystem explicit
The converters package has a dependency on the lfs module. See 65f39d49
for how it uses it. I realize this will only come up for people that use
that package, but I think it should be expected that the install
process will get everything setup to use any of the core packages that
ship with SILE at will. This lua library seems to be pretty widely used
so I don't see a problem with calling it out as a dependency. This is
something that package managers especially need to be made aware of.