use Noto Naskh for test and add missing test output
Adobe Naskh is a lot prettier, esp with mixed Arabic RTL text and LTR
numerals (which seems to be the point of this test), but it's non-free
and hard to come by. The Noto series on the other hand is easier to
install on Travis for the test suite. I have a copy of both compared the
output to see that the it was laying out the same other than the
different font metrics.