add expected output for OSX only test
This was generated on an OSX machine with the TTC font in question, so
it should be the _correct_ output. Note I was unable to duplicate the
fail state on this because the TTC loader only wants to work for me.
Also note the test that checks this needs to be skiped if not on OSX.