reverse build matrix so most wanted tests happen first
Although it works in parallel, Travis launches jobs from the top down
in the matrix. During busy times and if availability is low this can
mean the early ones on the list finish some time before the others.
Since I'm most interested in seeing the tests run in the latest versions
and can wait to hear about regression issues later, sort the matrix in
reverse order with the newest versions on top.