install SIL fonts from their own PPA
The Ubuntu 12.04 LTS repos have very very old versions of the SIL fonts
(e.g. 1.001 from 2009 instead of 2.020 from 2014) and the metrics have
gotten changed quite a bit, especially for the Arabic Scheherazade font,
but also lesser so for some of the others. This makes the regression
tests hard to run because it's not the font version any developer is
likely running on their system. Rather than benchmark the old ones, this
will just install the newest available versions into the Travis test
environment.

(Note mkdir is not necessary for the manual font install since unzip
 will create up to 1 directory on extract.)