manually install newer Linux Libertine
This is a package from Trusty because the one in Precise repos is way
out of date and has some bad metrics, ergo killing some of the
regression tests that depend on it. Since there isn't much in the
package but the font the deb file installs fine without much help.