remove benchmark output for examples directory
Not that we want to go breaking examples and such, but these will need a
different testing system than the ones we run using attack.pl triggered
by the existence of .expected files.

Here we actually have the PDF's in the repository anyway. If there _are_
regression tests that need to be run on the content of these examples it
should be moved into tests/.