separate compile tests from regression tests
The logic of checking whether all the stuff compiles vs whether it
exactly matches the reference implementation was getting too convoluted.
This separates the concerns so that only documentation and examples get
compile checks while the tests get compile and regression checks
(depending of the mode of the flag at run time). Travis for example runs
both modes, but separately so what failed is more apparent than if they
were lumped in together.