add minimal test case for Graphite issue #124
This test will only pass if either graphite support is not present or if
it is present and working. If it's present and not working it will fail.

Note this is different than the original test because looking at just
how ألمّ was typeset is only evident in a visual comparison of the PDFs.
As the overall word token is in the same positions it didn't register
in the debug output. This variant has another instance of the same
problem that affects stacked diacritic marks. As whether these get
properly combined or not affects overall line height calculations, this
will show up on the debug output.