rewrite SILE vs InDesign section for easy reading
Note removal of `\supereject` hack as the page break no longer falls in
the same place so the hack no longer solves the problem it was meant to.
I'm leaving off any sort of fix for the issue as a real fix in the
layout engine is on the TODO list for the next version.