add Travis test run without graphite support
I didn't bother adding this for all versions of Lua because I don't know
of any specific version paring problems. The known Graphite related
issue succeeds or fails in exactly the same way on 5.3 and 5.2 and
across platforms (hey consistent tests are a good thing!). On the other
hand the most likely systems to _not_ have support are older ones, so in
picking a version to test without it I went with 5.2.