switch to Lateef to avoid Graphite issues
The Scheherazade font triggers issue with Graphite + Harfbuzz that cause
this test to fail even if the script stuff is working. Lateef on the
other hand has support for the scripts being tested and will properly
test that feature without triggering the other issue (which is being
tested for in tests/bug-124.sil anyway).