print version to STDERR instead of STDOUT
Printing the version number to STDOUT makes it hard to adapt to assorted
ways a piped IO stream might be used. A classic case in point is our
very own regression tests. The *.expected output files should not have
version strings in them, just the layout data. Outputing this info to
STDERR still gives the user the reminder and piece of mind of hearing
from the core app as it starts running without leaving anything in the
stream that would get picked up by a pipe.