Cleanup output for easier test parsing
First of all this removes the output file name from the debug output. I
can't think that this is actually _testing_ anything useful such as not
being able to write to a file (which would be a STDERR message and an
abort anyway) and having the filename as part of the debug output
hinders being able to script the testing process. Using the `-o` flag I
can send the output various places (such as marking the filename with
the commit SHA) that can then be compared in various ways, but if the
debug output has a hard coded path this adds hassle to the process.

Additionally this puts SILE's main loop output of the input filename on
its own line for similar reasons. Not having a trailing newline on this
meant it was merged in the debug output with whatever came next, again
making diffs less informative.