Use font metrics from Travis for expected test output
An upstream issues in Harfbuzz causes different metrics to be returned
on different systems. Per issue #164 this sets the expected output of
the test to be what the Travis VM ends up reporting. This well allow the
test suite to pass (this test) until such a time as
`bh_ot_get_glyph_extents` is implemented and the test starts returning
consistent results across systems.