Add minlevel option to truncate counter output
The maxlevel is nice if you just want an upper division (theChapter
instead of theChapter + theSection) but what if you want just
theSection? Or in a more common scenario if you have parts, outputting a
counter that has the chapter number without outputting the part number
is normal.

Usage to output just the last level of the counter:

    SILE.call("show-multilevel-counter", {minlevel = options.level})