documentation/sile.sil: small fixes
Those are attempts at some small fixes I tried to do while reading the manual. I'm not sure if all of them are right, so please review.

Especially, I'm not sure how `SILE.Length` (or `SILE.length`?) and `Glue` (or `SILE.Glue`? `SILE.newGlue`?) are intendet to be named on lines ~1677-1678 (even after a cursory look in the code). But naming one in a different way than the other looks inconsequential.