Suppress 2× repetitions of leader when there is only space for 1×
The previous code would output a minimum of two repetitions of the
leader even after calculating that there was only space for one. This
would push the page number out of bounds in TOC entries that filled up
most of the space. The calculation for 0× repetitions worked and
anything over 2× worked, but if there was really only space for 1×
filler it was getting output twice anyway.