Make footnote font configurable
Previous code had a hard coded way of setting the foootnote font size to
9pt. The \footnote:atstart command could be fired up to override this,
but that fires only after the value has been set once. Additionally this
makes it hard for a class to customize and still make it easy for the
user to over-ride. A class should be able to set the default font size
while still letting the user run code of their own at start (without
having to manually include the font settings if they don't want to
override them.