Fix base url so relative links work
Fixes #294

This does _not_ allow the site to be hosted under forks on Github Pages
as it should because Jekyll there does not set the value correctly, but
it does allow a local Jekyll instance to be hosted under a subdirectory
if you can set the `--baseurl` flag. There are some other issues with
that, but this should restore the main site to full functionality.