Don't reset spaceskip inside \justified
This is a revert of what seems to be a stay setting added in 20d0617.
Reseting the spaceskip value to the font default value for just one
environment but not all the similar ones is inconsistent at best and I
can't figure out what problem it solves. Either this should be done in
all the justification environments so that users come to expect the
value to be changed or (as in this commit) it should be left alone and
whatever gets set in the parent node should carry through to justified
environments.