Add \strong for ease of document format conversion
Converting documents between formats (e.g. with Pandoc) is a kettle of
fish, but having a few standard markup tags available that have obvious
correlations in other markup languages makes the process a lot easier.
The name of `strong` was chosen to pair with our existing `em` also used
in HTML (as opposed to `textbf` which might have paired with `textit` if
we were using LaTeX as a reference).