Pass objects rather than strings through to TOC
With the previous arrangement only a simple string could be passed along
as a TOC entry. If it had any formatting codes inside the item (e.g. an
object with a string, a line break command, and another string) only the
first bit of string would get passed along. With this arrangement
whatever you send over comes out. This allows for more flexible TOC
entry typesetting where a straight string might not be all the
information coming over from the chapter title.