Fix broken test for issue #264
The previous value of 420pt was no longer enough to trigger this bug.
Reverting the code change in fd4ee10 no longer triggered a problem
because no page break was being encountered. 1 more pt would have fixed
this to trigger the old bug (and hence show whether the fix worked) or
not, but I took the opportunity to standardize on A5 paper size and
adapted the test accordingly.