Normalize test for issue #283 to work on Travis
Travis doesn't (yet) have the Libertinus fork version of the font used
in this test. Also in order to trigger the widest range of failures
possible the original test used a very specific page and font size
combination. This adaption allows it to work on A5 paper but setting the
corresponding font size that trips up the most break points.