Update expected output of test for issue #79
Note the page and line metrics changed a little so the break point on
this test was no longer falling on the hyphenation point that was
trouble. This has been tweaked to re-trigger the bug if not patched.

Also note that due to other unrelated changes to the way the Unicode is
handled and the tokenizer works, this bug was very difficult to
re-trigger. Reverting the original didn't cut it, I also had to back out
a number of other recent changes and create a bastardized version of the
core. Don't expect `git bisect` to be useful if ever tracking this down.