Remove extraneous (and broken) breaks
Having the first of these in here causes a frame-break anyway because of
the flow order. Then the \framebreak makes another one, making the first
column blank and moving everything into the second column. After that
the bottom box content comes so late that it's on the next page, where
the frame doesn't even exist.