Fix Lua syntax faux pas and optimize table inserts
The return value of table.insert() is true for insertions into a table
with length greater than one, but 0 for inserting the first item into a
table. This is not a useful test for these functions. The only failure
case I know of would raise a Lua error and isn't something we would fix
based on the return of these push functions anyway.

As long as we're fixing use of table.insert() AND have to return
something else anyway, just drop it in favor of a faster syntax.