Keep track of margins per vbox in the output queue
Fixes #262

This is approach is roughly (but not exactly) what was discussed on the
relevant issue. Unfortunately keeping a node in the queue that sets
margins isn't very useful because we cannot actually set those values
when we run across the node, but we may need to reset them later on
material from the same paragraph.

This solution leverages the fact that the rskip value doesn't change
during normal typesetting operations inside of a paragraph (whatever
the value is or however many times it's changed it doesn't stick until
you leave horizontal mode). Knowing this, we tag the vbox in the output
queue with its margin information every time we switch to vertical mode.

Then in the pushback() function when we re-process the nodes, each line
comes across in a separate vbox with information about its margins per
its original context. We replay these  values on the current typesetter
as we stuff it back into the queue.