Refine pushBack() test case to use varying width frames
See #262

The previous page based test had identical frame sizes between page
breaks so it didn't really test the hard case. The framesets here at set
to trigger the bug with the same content as the page based test so if
need for testing one can comment out the frameset and test this across a
page break.