Always inhibit leading after explicit VGlue
Fixes #344

We need to either do this all the time or none of the time. Previously
SILE was doing this for explicit vglue in the pushBack() routine but not
for normal pushes. We could not do this at all and expect users to deal
with the existing line leading if they use manual skips, but that
doesn't seem like the most generally useful approach. If we _were_ to go
that route, then it shouldn't be done in the pushBack() routine either.

This solution leaves one live of code that fixes it for both places
since it is always done in the pushExplicitVglue() function.