Mark some hfillGlue nodes as non-discardable
Fixes #301

Mark some hfill glues as an item that should not be discarded when
tearing apart and rebuilding vboxes and check for that property in the
pushBack() routine before discarding it.

The same hfillGlue objects are being used for lskip/rskip values in the
ragged environments and elsewhere. One object or push method just won't
suffice. Sometimes (as in when used for ragged margins) it needs to be
discardable, other times (as in when called manually to fill the middle
of a line) it needs to be explicit.

Likewise parfillskip is not discardabale in the way that lskip/rkip are,
so save that too.

I chose to make the distinction when pushing it to the output queue,
although nodes with different propertiens could also be produced.