Revert to SILE's previous default leading behaviour
See #344

It seems a lot of examples really do assume that manual skips will be
added to the normal leading. This seems unintuitive to me, but given how
many places it is already in effect, that change should probably be a
separate issue. If nothing else the test suite will be a lot easier to
fix with this in a separate branch.