Rewrite pushBack() again to address logic bugs
See #262

The previous iterations re-broke the test for #61 -- having never
really been fixed. The loop to drop penalties was alright as originally
created to fix that issue but it didn't play nice with other situations.

Meanwhile the main loop in pushBack() had such convoluted if statements
that logic bugs that should have been obvious kept slipping through.

Combining these loops and reversing the if logic to test for things to
drop first then keeping everything else should make it much easier to
maintain and catch bugs.

I'll confess I don't understand this the logic on discarding
discretionaries, changing it breaks everything (as show in the test for
269). Without keeping the first discretionary the line doesn't wrap at
all and overflows the margin, then dumps two leadings at once.  Without
marking it as unused it produces two hyphenations an the wrap point.