Update expected test results after pusBack() rewrite
All of these have been reviewed by hand to see what actually happened to
the output. Almost all of them were symptomatic of bugs in pushBack()
that just weren't related to the main point of the test. A few of them
had minor metric changes due to the change in when leading  is
calculated in relation to skips.

The remaining failing tests are failing legitimately (something is still
wrong with switching frames).