Only run coverage tests if regression tests pass
This will cut the time to fail Travis test runs to 1/3 its current time
(from ~12min to ~4min) at the cost of taking a few extra minutes to give
a full green light because it has to re-run the regression tests with
coverage enabled. This extra time is mitigated that the non-coverage
tests will also pass in a much shorter time so impatient people watching
the test run will see it's going to succeed without waiting for the
final spurt.