Run regression tests before building examples
The test suite gives much more specific results and is usually what
we're waiting around for in a test run. Do that part first so that we
can fail quicker in the event of an issue before trying to build the
documentation and other examples. We still want to test the latter, but
want to fail quicker if some specific issue is found.