Re-order build matrix for speedier results
Our one test with coverage enabled takes longer that the others to run.
When Travis load is low it starts all the tests at once, so this won't
matter at all. However when the load is high it starts a maximum of 4
jobs per repository (or possibly even less). This means that the first 4
items in the matrix run and the last ones don't even start until they
finish. Putting the long running job late in the matrix delays the
overall time under peak conditions. Firing it off first cuts 4-6 minutes
off the overall test run time if we are hit a low maximum job limit.