Replace pairs() and table.insert with faster alternatives
Note that the scope of control variables set by the for control
structure is always local to the loop so there is no need to declare the
variables as local in the parent scope. In fact doing so introduces
possible confusion by suggesting such variables would even be available
at the end of the loop, which they are not.