Allow discardable skips
The current \skip command and friends always push non-discardable
glues to the stack. There are actually a few cases where explicitly
pushing a discardable glue would be useful. This can be done by calling
the corresponding typesetter function directly, but the usage is less
intuitive that the existing skip function. This adapts the existing
function to simply accept an optional additional argument and work for
both scenarios.