Correct typos, note libtool dependency
Fixes #376

I don't know where it came from but there was a bogus null character in
the documentation as well as another typo. This fixes those text as well
as adding libtool to the list of dependencies for RPM based distros.
Most of them actually include this as a dependency of other build tools,
but some recent versions of Fedora have dropped it, so being explicit
might help some.