Generalize measureDim() function to measureChar()
The usage for the original function was non intuitive because it
returned a different axis depending on what letter you sent. If you
_knew_ that 'x' was handled differently than 'n' it could be used, but I
got myself in trouble assuming from the context that I knew what it did.
This fixes the name and return values to be the same in all contexts. In
other words you can read the code and know what it does without knowing
the input values.