Allow classes to be loaded from specific paths
The previous code required custom classes to be at a hard coded path
relative to the search path (SILE_PATH and/or package.path). This
extends the mechanism already used to find files to be used for classes
as well with a preference for a Lua file prefixed with "classes". Baring
that existing the usual litany of paths is checked.