Return warning rather than error on frame missing
Scripting for whether we have a frame to send things to is hard (i.e.
you have to write your own replacement for this) if it doesn't return
anything. This enables something like:

    local myFrame = SILE.getFrame("myFrame")
    if myFrame then SILE.typesetNaturally(myFrame, function() ... end) end

Additionally the variable check there is silly because with Lua's
variable scoping it should never get called. Now if for some reason we
get _passed_ a frame, then I can see just giving it back.