Refactor language command to always take the same args
The previous usage was inconsistent with the way most other commands
work in SILE. Previously this took one argument in the event of no
content and a different  one (to do the same function) if there was
content. We already check for content or not, so just behave like most
other functions applying the setting globally if no content and wrapped
in a temporarily environment if there is content.