Tweak debug() to accept variable number or args
Previously if you wanted to pass more than one item to this function you
had to wrap it in a table and pass it as a single argument. Lua actually
has a way to handle a variable number of arguments passed to a function.
This won't break the previous usage because a table will still show up
as a single argument, but it will make the usage more versatile because
if you _want_ you can tack on more arguments to a call without having to
go back and wrap it it a table.