Add utility function to dump Lua tables
The debug function is nice for getting SILE's thoughts about what
something is, but because it tries to convert everything to strings it's
not great for figuring out Lua's take on things.  This function
pretty-prints one or many tables or other values in a way that's human
readable and gives a better view into the Lua data structures.

This requires an extra Lua library some systems won't have but my
thought was it isn't necessary to include this as a SILE dependency as
this function is really only for debugging SILE itself or for people
really hacking on their Lua code. It shouldn't be required in every day
SILE usage and I wouldn't leave and dump() calls lying around in the
source so it shouldn't break anything to have it here without calling
for an extra dependency. Developers that want to use this can fetch the
Lua rock.

Dumping this to STDERR means this can be split this out from SILE's
other output.