Adapt debug() and dump() functions to unwrap single items
This allows single items passed (i.e. now wrapped in a table) to be
printed without being wrapped in an extra table. The end result should
now be identical to previous iterations of debug() when passing a single
argument and only show a series of arguments as table if more than one
is actually sent.