Handle input files as named pipes
This allows using STDIN or named pipes as input files. The disadvantage
of this implementation is that it reads the entire contents into memory
even if it's going to reject it based on sniffing the first few bytes of
content later. The advantage is you  can do all sorts of Unix things
with files that the in-file seek was blocking.