Add \set-default command to change the default value of the setting.
One of the user case is to replace the default font of the document.
Since the default font is hard-coded into the script. Everytime when
execute SILE.settings.reset, we will go back to that font. This behavior
is not desired when you are using book class with footnotes and folio.