refactor(languages): Tidy up ICU-based shaper
The best way to solve the French bug is to create a French-specific nodemaker. But currently writing a nodemaker involves copying a lot of boilerplate code. Some of this is also confusing because there is a lot of overlap between the ICU nodemaker (which asks ICU about breakpoints) and the “general Unicode” nodemaker (the `dealWith` method, which deals with tokens in isolation). Hopefully this refactoring will help with the mess.

The next stage is tests for and refactoring of languages which use their own nodemakers, and then we can get on to French.