refactor!(classes): Rename :pre/:post functions to :prenumber/postnumber
BREAKING CHANGE: The :pre and :post function names didn't make any sense
because they were not actually before or after the chapter (or whatever
level) headings but just around the numbers. This makes it a lot more
clear in say a book or project library what is actually being done.

This is a step towards making the heading levels easier to style.