feat(backends): Add dummy output backend for dry runs
This is useful for when you don't want to produce any document at all
but only generate side effect files such as make dependencies. There
isn't a huge speed savings because all the shaping and layout commands
are still being run (this could be improved) but at least it provides
a mechanism not to generate any output files.