refactor(shaper): Use OT metric functions in Harfbuzz where available
Harfbuzz 1.0.0 included its own way of measuring font metrics, but did not include support for CFF-based fonts. So we used Harfbuzz OT functions for Truetype fonts (which were guaranteed to be have a `glyf` table rather than a `CFF ` table), and asked HB to call out to Freetype to measure OpenType fonts.

Now HB has its own metrics functions for both CFF and glyf fonts, so we should use them where available. Note this will fallback to Freetype for older Harfbuzz versions. This matters because the measurement of the vertical metrics differ in the case of CFF fonts between Freetype and Harfbuzz. See #639.