refactor!(core): Use Lua's built in resolver instead of our monkey work
BREAKING CHANGE: This removes the auto-guessing file extension
mechanism that allowed *.sil files to be loaded without specifying the
full file name with extensions. A command like `sile test` will no
longer find and build sile.sil, you must run `sile test.sil`. The
mechanism that was doing this was a hack than only worked in some
scenarios anyway, and dropping it instead of trying to cover all the
edge cases will make it that much easier to use and document.
Importantly it avoids edge cases where both *.xml, *.sil, and/or *.lua
files all have the same name and the loader really has so idea which one
you mean.

Note that _packages_ may still be loaded with no file extension, this
does not affect the `require()` mechanism that looks for *.lua and
various other incantations by default.