chore(build): Explicitly bundle specific Lua Rock versions
These are all the current latest versions of the rocks in question. We
don't actually need exactly these versions (as far as we know) nor are
they the minimum versions required. Distro packages are free to test
against older or newer versions using `make test` against whatever
versions they have, but these are what we are feeding into our CI for
testing and eventually bundling with source tarballs and any packagers
that require support bundling. The main point of spelling out the
versions here is so that our testing is consistent, rebuilds of previous
releases always get the matching dependency code, and future problem
tracing will be easier with `git bisect` available to trace where
problems were introduced.