fix(core): Restructure Lua path order to only prefix default
It turns out the last rework of this was a bit too aggressive. First we
started replacing the path entirely, then we added the default system
back back on the end, then we started prefixing instead of replacing it.
As a result the system default paths were added twice.

Second we were using autoconf to tell us what the system path was at
build time instead of letting the Lua interpreter tell us that at run
time. This was causing weird results it testing packaged versions.

Lastly, autoconf usage was just a mess.

This tries to only prefix things we know we need onto the default Lua
runtime values.