refactor(core): Rename typesetter.pageBuilder() to <>.buildPage()
We also have something called pagebuilder which is a toolset for
building pages. Using the same name for different things was confusing,
this makes it quite a bit more apparent what is actually doing what.
Ideally this should probably be completely refactored so that
typestetter doesn't even have this logic at all and the actual build
step is a function of the pagebuilder as well.