fix(measurements): Actually deprecate old constructors
This should have been done when the code was reorganized, but it was
hard enough to test everything at once without warnings cluttering up
the output, so I waited and then forgot about it. The shims were already
in place, all this does is actually throw warnings on the deprecated
usages.