feat(packages): OpenType justification package
The OpenType JSTF table allows a font to communicate information to a layout system about how to improve justification. Nothing implements it yet, so we’re going to be the first.

For now, we use data within the font to set the spacing stretch and shrink. We look up the current script/language combination in the JSTF table and see if any of the “justification suggestions” give positioning information for the space glyph. If so, we use this information to make space glue.