ci(travis): Only test build finalized braches and PRs
Lots of our Travis jobs have been failing because we're causing Travis
to be IP blocked by people that serve font files or other code we
download. Besides, it often takes huge amounts of time to catch up and
get test results after pushing a few feature branches around. This
should cut down on the things we ask Travis to churn through by limiting
it to master & tagged releases plus PRs. If we open PRs we want those
checked of course, but not every branch we throw around.