fix(build): Touch Makefile.in to avoid automake errors
If for some reason Makefile.in is older than acloacl.a4 and other
automake generated files, it will want to regenerate it ... which
requires automake tools, and specifically wants matching versions.

This simple action (after we're done mucking around with the other
files to remove automake's handling of "core") keeps it from thinking it
needs to rebuild anything.