ci(actions): Placate old automake and Docker weirdness
we have a bug somewhere either in automake itself or in our dependencies
getting from zero to dist. It works on 1.16.2, but in Docker with 1.16.1
it fails. The two options are either to drop the -j arg and run a single
thread or make before make dist. The latter being faster and also more
granular in the output, that's what we're up to.