fix(classes): Reset parindent's inside \center command
This will cause a lot of documents to render differently, but I think
it's the right thing to do. When you ask for something to be centered
and it ends up shifted off to the side, that feels wrong. Inhibiting the
document's normal parindent entirely for the duration of the \center
seems like the best solution.

Also now warns if used and there already nodes in the queue. This should
at least tip people off that there is a problem with their document.