ci(travis): Drop docs & examples from Travis tests
The documentation in particular takes a _long_ time to compile and
doesn't (or shouldn't) add much to the coverage of other tests. In
particular I don't think we need to run those across the whole matrix of
Lua and HarfBuzz versions.

Note a pass/fail grade for building the docs & examples is still
a requirement for merging, but it is done as part of the Github Actions
workflow that also posts artifacts of those PDFs.