chore(build): Use Github download for Symbola font
The download from Archive.org is still valid (and even still active in
this makefile) but it frequently gets blocked when we do too much CI
testing in quick succession on branches that are failing (and hence
don't cache fonts). This shortcut is a more precise rule (so takes
precedence) and just uses some random dev's Github repo cache of the
font file itself. This should be a bit more robust. Commenting out this
new rule should revert to the other download if needed.