ci(travis): Revert double fix that broke regular Lua
Fixing the same problem two ways makes it broken again. We can't both
use full version paths *and* tell the build system to only install
non version suffixed binaries. That worked for LuaJIT because the
Makefile tries to do both anyway, but for regular Lua we need one
solution or the other.