refactor(packages): Rename \include-svg-file to \svg
This is more consistent with \img, and we really don't need the extra
command name complexity. I expect to add inline SVG capability to this
someday just like \script, but it has worked out just fine to use that
for includes (with a src= option) and for inline content (without but
with passed content). I don't see why we can't do the same thing for
this.